pyproject.toml
[tool.poetry]
name = "Chatterbox voice and text assistant"
version = "0.1.0"
description = "Desktop applications"
authors = ["Viachaslau_Kazakou <sly.kazakoff@gmail.com>"]
packages = [{include = "src"}]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pyinstaller = "^6.12.0"
pytest = "^8.3.4"
black = "^25.1.0"
flake8 = "^7.1.2"
flake8-pyproject = "^1.2.3"
isort = "^6.0.1"
pyqt6 = "^6.8.1"
sounddevice = "^0.5.1"
numpy = ">=1.20.0,<2.0.0"
torch = "2.2.2"
transformers = "^4.50.3"
accelerate = "^1.6.0"
openai-whisper = "20240930"
ollama = "^0.4.7"
pygments = "^2.19.1"
myst-parser = "^3.0.0"
jupyter-book = "^1.0.4.post1"
matplotlib = "^3.10.1"
roman-numerals-py = "3.1.0" 
# sphinx = "8.2.3" 
sphinx_autodoc_typehints = "1.24.0"
langchain = "0.1.10"
librosa = "^0.11.0"
sentence-transformers = "^5.0.0"
faiss-cpu = "^1.11.0"
hf-xet = "^1.1.5"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py312']

[tool.flake8]
max-line-length = 120
extend-ignore = "E203, W503"

[tool.isort]
profile = "black"

